+++ CHANGELOG.md
@@ -9,8 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added Stimulus JS Controllers
   [`auto-play-audio`, `autocomplete-address`, `checkbox-toggle`, `elements-overlap`,
-  `focus-on-connect`, `input-on-change`, `print`, `radio-toggle`, `submit-button`,
-  `submit-on-change`].
+  `focus-on-connect`, `input-on-change`, `print`, `radio-toggle`, `submit-button`].
 
 ## [0.12.0] - 2022-06-24